requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
ollama>=0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pytz==2024.2
